fix: playlist details page
