+++ src/pages/[language]/library/playlists/[id].ts
@@ -1,7 +1,7 @@
 import { DehydratedState } from '@tanstack/react-query';
 import { GetServerSidePropsContext, GetServerSidePropsResult } from 'next';
 
-import LibraryPlaylistDetail from '~containers/library/playlist/detail';
+import LibraryPlaylistDetail from '~containers/library/playlist/libraryPlaylistDetail';
 import { storeRequest } from '~lib/api/storeRequest';
 import { prefetchQueries } from '~src/__generated__/prefetch';
 import getDehydratedProps from '~src/lib/getDehydratedProps';
