+++ src/pages/[language]/playlists/[playlist].ts
@@ -1,12 +1,12 @@
 import { GetServerSidePropsContext, GetServerSidePropsResult } from 'next';
 
 import { getPlaylistPageData } from '~containers/library/playlist/__generated__/detail';
-import LibraryPlaylistDetail, {
-	ILibraryPlaylistDetailProps,
+import PlaylistDetail, {
+	IPlaylistDetailProps,
 } from '~containers/library/playlist/detail';
 import { storeRequest } from '~lib/api/storeRequest';
 
-export default LibraryPlaylistDetail;
+export default PlaylistDetail;
 
 export async function getServerSideProps({
 	req,
@@ -14,7 +14,7 @@ export async function getServerSideProps({
 }: GetServerSidePropsContext<{
 	language: string;
 	playlist: string;
-}>): Promise<GetServerSidePropsResult<ILibraryPlaylistDetailProps>> {
+}>): Promise<GetServerSidePropsResult<IPlaylistDetailProps>> {
 	storeRequest(req);
 	const { playlist } = params?.playlist
 		? await getPlaylistPageData({
